refactor(models): 重构工具类并优化表情包相关功能
- 新增 Tools 类，整合 meme 数据处理和网络请求相关方法
- 重构 meme.js，使用新的 Tools 类处理数据和请求
- 更新 update.js 和 checkRepo.js，引入新的 Tools 类替换 Utils
- 优化 meme 数据下载和生成逻辑，支持强制更新功能
- 新增关键词转键值和黑名单检查方法
