+++ models/index.js
@@ -3,6 +3,7 @@ import Meme from './meme.js'
 import Utils from './utils.js'
 import Rule from './rule.js'
 import Args from './args.js'
+import Tools from './tools.js'
 import checkRepo from './checkRepo.js'
 
-export { Request, Meme, Utils, Rule, Args, checkRepo }
+export { Request, Meme, Utils, Rule, Args, Tools, checkRepo }
